requirements.txt
docker_py==1.10.6
Flask==2.0.2
kubernetes==24.2.0
matplotlib==3.4.3
Mesa==1.0.0
numpy==1.21.3
pymoo==0.6.0
requests_futures==1.0.0
